LayoutTests/fast/dom/document-all-select.html

<html>
<head>
<script src="../js/resources/js-test-pre.js"></script>
<script>
if (window.testRunner)
    testRunner.dumpAsText();
</script>
</head>
<body>
<select name="myselect" style="visibility: hidden;">
<option value="1">A</option>
<option value="2">B</option>
</select>
<script>
description("This tests that &lt;select&gt; elements are available by name through document.all.");
shouldBe("document.all.myselect.length", "2");
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
