safari-600.7-branch/LayoutTests/fast/dom/document-all-select-expected.txt

This tests that <select> elements are available by name through document.all.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.all.myselect.length is 2
PASS successfullyParsed is true

TEST COMPLETE
